append TWITTER to env, anon receiver, start implementing get users
