+++ .env.sample
@@ -1,5 +1,4 @@
-API_KEY=api_key from your twitter app
-API_SECRET=api secret from your twitter app
-AUTH_CALLBACK=callback you will use for client application (e.g. `http://localhost:3000/callback`)
-OAUTH_TOKEN=access token from oauth protocol or twitter developer account page
-OAUTH_TOKEN_SECRET=access secret from oauth protocol or twitter developer account page
+TWITTER_API_KEY=api_key from your twitter app
+TWITTER_API_SECRET=api secret from your twitter app
+TWITTER_OAUTH_TOKEN=access token from oauth protocol or twitter developer account page
+TWITTER_OAUTH_TOKEN_SECRET=access secret from oauth protocol or twitter developer account page
