+++ README.md
@@ -20,4 +20,4 @@ In order to use many of the features of this application, you must provide api k
 
 4) Create a file named `.env` in the root directory of this application following the example of `.env.sample`. Fill in the API key and Secret generated in the previous step in the `.env` file.
 
-5) The API Key and Secret are used to identify your app. The quickest and easiest way to authenticate requests to the twitter api from within this client would be to generate an Access Token and Secret within the twitter developer portal. Under your application name, click `Keys and tokens` then click `Generate` next to `Access Token and Secret`. Copy these to your .env file under `OAUTH_TOKEN` and `OAUTH_SECRET`. You are now ready to make requests to the twitter api!
+5) The API Key and Secret are used to identify your app. The quickest and easiest way to authenticate requests to the twitter api from within this client would be to generate an Access Token and Secret within the twitter developer portal. Under your application name, click `Keys and tokens` then click `Generate` next to `Access Token and Secret`. Copy these to your .env file under `TWITTER_OAUTH_TOKEN` and `TWITTER_OAUTH_SECRET`. You are now ready to make requests to the twitter api!
