+++ internal/api/tweets.go
@@ -6,18 +6,12 @@ import (
 	"gotwitter/internal/types"
 )
 
-type GetTweetsOptions struct {
-	Ids         []string
-	Expansions  []string
-	TweetFields []string
-}
-
 type Tweets struct {
 }
 
 // https://developer.twitter.com/en/docs/twitter-api/tweets/lookup/api-reference/get-tweets
 //   Parameters:
-//     `ids` []string - array of ids to query **REQUIRED**
+//     `ids` []string - array of tweet ids to query **REQUIRED**
 //     `expansions` []string - array of expansions (see link for available expansions)
 //     `media.fields` []string - array of media fields to include. Requires `attachments.media_keys` expansion
 //     `place.fields` []string - array of place fields to include. Requires `geo.place_id` expansion
@@ -25,7 +19,7 @@ type Tweets struct {
 //     `tweet.fields` []string - array of tweet fields to include.
 //     `user.fields`  []string - array of user fields to include. Requires certain expansions (see link)
 //
-func (t *Tweets) Get(options types.GOTOptions) ([]types.TweetData, []types.Error) {
+func (*Tweets) Get(options types.GOTOptions) ([]types.TweetData, []types.Error) {
 	response, errors := network.Get[types.TweetsResponse](endpoint.Tweets, options, []string{"ids"})
 	if errors != nil {
 		return []types.TweetData{}, errors
@@ -43,7 +37,7 @@ func (t *Tweets) Get(options types.GOTOptions) ([]types.TweetData, []types.Error
 //    `media.tagged_user_ids` []string - a list of user ids being tagged in the Tweet with media
 //    `poll.duration_minutes` int - Duration of the poll in minutes for a Tweet with a poll
 //
-func (t *Tweets) Create(payload types.GOTPayload) (types.CreateTweet, []types.Error) {
+func (*Tweets) Create(payload types.GOTPayload) (types.CreateTweet, []types.Error) {
 	response, errors := network.Post[types.CreateTweetResponse](endpoint.Tweets, payload)
 	if errors != nil {
 		return types.CreateTweet{}, errors
@@ -55,7 +49,7 @@ func (t *Tweets) Create(payload types.GOTPayload) (types.CreateTweet, []types.Er
 // Payload parameters:
 //    `id` string - the id of the tweet to be deleted
 //
-func (t *Tweets) Delete(payload types.GOTPayload) (types.DeleteTweet, []types.Error) {
+func (*Tweets) Delete(payload types.GOTPayload) (types.DeleteTweet, []types.Error) {
 	tweet_id := payload["id"]
 	response, errors := network.Delete[types.DeleteTweetResponse](endpoint.TweetById(tweet_id))
 	if errors != nil {
