+++ internal/client.go
@@ -4,9 +4,10 @@ import "gotwitter/internal/api"
 
 type GOTClient struct {
 	Tweets *api.Tweets
+	Users  *api.Users
 }
 
-func New( /* api keys/secrets, config, etc */ ) *GOTClient {
+func NewClient( /* api keys/secrets, config, etc */ ) *GOTClient {
 	tweets := &api.Tweets{}
 	client := GOTClient{Tweets: tweets}
 	return &client
